requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
firebase-admin==6.6.0
pycryptodome==3.21.0
jinja2==3.1.4
python-multipart==0.0.18
markdown==3.7
pytest==8.3.4
httpx==0.27.2
mnemonic==0.21
pytest-playwright==0.4.4
pytest-base-url==2.0.0

